refactor(realsense): the rust module is RealSenseCamera; python capture deleted
pyrealsense2-extended dropped with it; DepthCameraHardware stays for ZED and the sim.
