+++ dimos/hardware/sensors/camera/realsense/README.md
@@ -1,9 +1,9 @@
-# RealSense SDK Install
+# RealSense
 
-1) Install the Intel RealSense SDK:
-   - https://github.com/IntelRealSense/librealsense
+Capture is the rust module in `rust/`, built through its own flake (librealsense2 is not in the root shell):
 
-2) Install the Python bindings:
-   ```bash
-   pip install pyrealsense2
-   ```
+```bash
+cd dimos/hardware/sensors/camera/realsense/rust && nix develop path:. -c cargo build --release
+```
+
+`camera.py` launches it; `dimos run real-sense-camera-vis` shows color, depth and the cloud in Rerun.
