+++ dimos/robot/all_blueprints.py
@@ -86,7 +86,6 @@ all_blueprints = {
     "mid360-realsense-record-with-pcap": "dimos.robot.assembly.mid360_realsense_30:mid360_realsense_record_with_pcap",
     "openarm-planner-coordinator": "dimos.robot.manipulators.openarm.blueprints.basic:openarm_planner_coordinator",
     "openyam-planner-coordinator": "dimos.robot.manipulators.openyam.blueprints.basic:openyam_planner_coordinator",
-    "real-sense-camera-native-vis": "dimos.hardware.sensors.camera.realsense.blueprints:real_sense_camera_native_vis",
     "real-sense-camera-vis": "dimos.hardware.sensors.camera.realsense.blueprints:real_sense_camera_vis",
     "spot": "dimos.experimental.robot.bosdyn.spot.blueprints.spot:spot",
     "spot-record": "dimos.experimental.robot.bosdyn.spot.blueprints.spot_record:spot_record",
@@ -269,7 +268,6 @@ all_modules = {
     "quest-teleop-module": "dimos.teleop.quest.quest_teleop_module.QuestTeleopModule",
     "ray-tracing-voxel-map": "dimos.mapping.ray_tracing.module.RayTracingVoxelMap",
     "real-sense-camera": "dimos.hardware.sensors.camera.realsense.camera.RealSenseCamera",
-    "real-sense-camera-native": "dimos.hardware.sensors.camera.realsense.native.RealSenseCameraNative",
     "real-sense-mount-tf": "dimos.hardware.sensors.camera.realsense.blueprints.RealSenseMountTf",
     "receiver-module": "dimos.utils.demo_image_encoding.ReceiverModule",
     "recorder": "dimos.memory.module.Recorder",
